requirements.txt
ipykernel
pandas
numpy
matplotlib
seaborn
scikit-learn
python-box
pyYAML
mlflow
dagshub
-e .